Organized create new trip files and project file/folder structure.
