+++ together-wherever/src/components/PlaceSearchBox.tsx
