+++ together-wherever/src/constants/create-new-trip.ts
@@ -1,0 +1,4 @@
+export const CREATE_NEW_TRIP_CONFIRMATION_DIALOG = {
+    topic: "Create New Trip",
+    desc: "Please confirm that you want to create a new trip."
+};
