+++ together-wherever/src/constants/vote.ts
