+++ together-wherever/src/fetcher/create-new-trip.ts
@@ -1,8 +1,30 @@
 'use client';
 
 import axios from "axios";
-import { CreateNewTripBodyInterface } from "@/app/utils/types";
+import { CreateNewTripBodyInterface } from "@/utils/types";
+import { mockUsersData } from "@/mocks/create-new-trip";
 
+// simulate fetching users data
+export const fetchingUsersData = () => {
+    try {
+
+        // const response = await axios.get('/api/getUsersData');
+        // if (response.status === 200) {
+        //     const { usersData } = response.data;
+        //     return usersData;
+        // } else {
+        //     throw new Error(`Unexpected response status: ${response.status}`);
+        // }
+
+        const allUsersData = mockUsersData;
+
+        return allUsersData;
+
+    } catch (error) {
+        console.error("Error fetching voting page data:", error);
+        throw error;
+    }
+};
 
 export const createNewTrip = async (body: CreateNewTripBodyInterface) => {
     try {
