+++ together-wherever/src/fetcher/get-place-details.ts
@@ -1,0 +1,12 @@
+import axios from 'axios';
+import { PlaceDetails } from '@/utils/types';
+
+export const fetchPlaceDetails = async (placeId: string): Promise<PlaceDetails> => {
+    try {
+        const res = await axios.get(`http://localhost:8000/api/discover-place-details/?dest_id=${placeId}`);
+        return res.data as PlaceDetails;
+    } catch (error) {
+        console.error('Failed to fetch place details', error);
+        throw new Error('Failed to fetch place details');
+    }
+};
